+++ lib/src/features/repository/history_screen.dart
@@ -858,7 +858,15 @@ class _HistoryScreenState extends State<HistoryScreen> {
           ContextActionRoute.checkoutRemote ||
           ContextActionRoute.deleteRemote ||
           ContextActionRoute.cherryPickRemote ||
-          ContextActionRoute.hostLink:
+          ContextActionRoute.hostLink ||
+          ContextActionRoute.openRepository ||
+          ContextActionRoute.activateRepository ||
+          ContextActionRoute.closeRepository ||
+          ContextActionRoute.closeOtherRepositories ||
+          ContextActionRoute.removeRecentRepository ||
+          ContextActionRoute.copyRepositoryPath ||
+          ContextActionRoute.revealRepository ||
+          ContextActionRoute.openNestedRepository:
         return;
     }
   }
@@ -1342,7 +1350,15 @@ class _HistoryScreenState extends State<HistoryScreen> {
           ContextActionRoute.checkoutRemote ||
           ContextActionRoute.deleteRemote ||
           ContextActionRoute.cherryPickRemote ||
-          ContextActionRoute.hostLink:
+          ContextActionRoute.hostLink ||
+          ContextActionRoute.openRepository ||
+          ContextActionRoute.activateRepository ||
+          ContextActionRoute.closeRepository ||
+          ContextActionRoute.closeOtherRepositories ||
+          ContextActionRoute.removeRecentRepository ||
+          ContextActionRoute.copyRepositoryPath ||
+          ContextActionRoute.revealRepository ||
+          ContextActionRoute.openNestedRepository:
         return;
     }
   }